New tone test functions
Function to generate tone to the speakers.
Function to translate a note to a frequency.
Function to send a note to speakers.
Function to translate a note to MIDI tone.
Function to read a text file and send notes within to speakers.
Additional test function to read contents of a MIDI file.